LayoutTests/platform/glib/http/tests/websocket/tests/hybi/compressed-control-frame-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://127.0.0.1:8880/websocket/tests/hybi/compressed-control-frame' failed: Received invalid WebSocket response from the server
Test whether a compressed control frame is rejected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
